+++ changelog.md
@@ -10,6 +10,7 @@
 - Updated error responses on all webapps and groot to always return a json with a standard error format
 - Fixed kamehouse-shell scripts permissions
 - Updated redirect modal with response code on unsuccessful authorization on mobile
+- Updated docker build/run scripts
 - Fixed bugs
 
 #### v8.16